+++ docs/source/changelog.rst
@@ -1072,6 +1072,17 @@ Changes
   - ``cairosvg==2.4.0``
   - ``pillow==6.1.0``
 
+- Context-sensitive help on the CRATE web site, via
+  :class:`crate_anon.common.constants.HelpUrl`.
+
+- Amended ``show_sitewide_queries.html`` to remove ``<form>`` children of
+  ``<tr>``; see
+
+  - https://stackoverflow.com/questions/7737163/form-within-table-row-tag
+  - https://stackoverflow.com/questions/1249688/html-is-it-possible-to-have-a-form-tag-in-each-table-row-in-a-xhtml-valid-way/16941843
+
+
+
 ===============================================================================
 
 .. rubric:: Footnotes